requirements.txt
websocket-client
pandas
ta
numpy
requests
